Smoother bucket and claw operation
Dead zone value of 10 seems too low for XBOX controller where an idle
joystick may outpuvalues as high as 60. We can get away with smaller
values for the motors.

The neutral joysticks' position still gives non-zero redouts. The dead
zone value is there to neutralise them. However, it would be better if
we can compensate them to some degree. The calibration mode can now be
activated and the user should jiggle the joysticks a bit and then let
them stay in the netral position for a few seconds. However, this feature
is of a dubious benefit because the values read from joysticks seem to
drift at rest.