No PWM without logging
The loop delay that yields the CPU to the PWM maintenance was outside
the loop and thus not having any effect. Without logging through the
serial output which itself presumably allows the maintenance thread to
kick in, the motors practically do not move.

Interestingly, it somewhat works if channel 8 and above are used. It
seems that ESP32Servo library designates channels 0-7 as low speed
software based and the rest as interrupt based. However, even the
interrupt based channels work better with the delay.